% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg-package.R
\docType{package}
\name{arg-package}
\alias{arg}
\alias{arg-package}
\title{arg: Clean, Simple Argument Checking}
\description{
Functions for checking arguments in R package. Uses a simple interface and produces clean, informative error messages using 'cli'.
}
\author{
\strong{Maintainer}: Noah Greifer \email{noah.greifer@gmail.com} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

}
\keyword{internal}
