% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_color.R
\name{arg_color}
\alias{arg_color}
\alias{arg_colors}
\alias{arg_colour}
\alias{arg_colours}
\title{Check Color Argument}
\usage{
arg_color(x, .arg = rlang::caller_arg(x), .msg = NULL, .call)

arg_colors(x, .arg = rlang::caller_arg(x), .msg = NULL, .call)

arg_colour(x, .arg = rlang::caller_arg(x), .msg = NULL, .call)

arg_colours(x, .arg = rlang::caller_arg(x), .msg = NULL, .call)
}
\arguments{
\item{x}{the argument to be checked}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}

\item{.call}{the execution environment of a currently running function, e.g. \code{.call = rlang::current_env()}. The corresponding function call is retrieved and mentioned in error messages as the source of the error. Passed to \code{\link[=err]{err()}}. Set to \code{NULL} to omit call information. The default is to search along the call stack for the first user-facing function in another package, if any.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument is a valid color. \code{arg_color()} additionally requires that a single color is supplied. British spellings are supported.
}
\details{
A color specification is considered valid if \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}} returns a matrix when applied to it.
}
\examples{
f <- function(z) {
  arg_colors(z)
}

try(f("red"))       # No error
try(f("redhehehe")) # Error: not a valid color name

try(f("#A4A473")) # No error
try(f("~A4A473")) # Error: not a valid hex code

try(f(15))  # No error
try(f(-15)) # Error: not a valid color number

try(f(list("red", 15)))  # No error
try(f(list("red", -15))) # Error
}
\seealso{
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}
}
