% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_length.R
\name{arg_length}
\alias{arg_length}
\title{Check Argument Length}
\usage{
arg_length(x, len = 1L, .arg = rlang::caller_arg(x), .msg = NULL)
}
\arguments{
\item{x}{the argument to be checked}

\item{len}{\code{integer}; the allowed length(s) of \code{x}. Default is 1.}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument has a specified length.
}
\details{
\code{len} can contain multiple allowed counts; an error will be thrown only if \code{length(x)} is not equal to any value of \code{len}.
}
\examples{
obj <- 1:4

try(arg_length(obj, 1))
try(arg_length(obj, 4))
try(arg_length(obj, c(1, 4)))

# These test the same thing:
try(arg_length(obj, c(0:3)))
try(when_not_null(obj,
                  arg_length(1:3)))
}
\seealso{
\code{\link[=length]{length()}}, \code{\link[=arg_non_null]{arg_non_null()}} to specifically test that an object's length is or is not 0.
}
