% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_logical.R
\name{arg_logical}
\alias{arg_logical}
\alias{arg_flag}
\title{Check Logical Argument}
\usage{
arg_logical(x, .arg = rlang::caller_arg(x), .msg = NULL)

arg_flag(x, .arg = rlang::caller_arg(x), .msg = NULL)
}
\arguments{
\item{x}{the argument to be checked}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument is a logical vector (\code{arg_logical()}) or a logical scalar (\code{arg_flag()}), i.e., a single logical value. Logical values include \code{TRUE} and \code{FALSE}.
}
\details{
\code{NA} values in \code{arg_flag()} will cause an error to be thrown.
}
\examples{
obj <- TRUE

try(arg_flag(obj))    # No error
try(arg_logical(obj)) # No error

obj <- c(TRUE, FALSE)

try(arg_flag(obj))    # Error: must be a scalar
try(arg_logical(obj)) # No error

obj <- 1L

try(arg_flag(obj))    # Error must be logical
try(arg_logical(obj)) # Error must be logical
}
\seealso{
\code{\link[=is.logical]{is.logical()}}, \code{\link[rlang:scalar-type-predicates]{rlang::is_bool()}}
}
