% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_supplied.R
\name{arg_supplied}
\alias{arg_supplied}
\title{Check Supplied Argument}
\usage{
arg_supplied(x, .arg = rlang::caller_arg(x), .msg = NULL, .call)
}
\arguments{
\item{x}{the argument to be checked}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}

\item{.call}{the execution environment of a currently running function, e.g. \code{.call = rlang::current_env()}. The corresponding function call is retrieved and mentioned in error messages as the source of the error. Passed to \code{\link[=err]{err()}}. Set to \code{NULL} to omit call information. The default is to search along the call stack for the first user-facing function in another package, if any.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument with no default value was supplied. An error will be thrown if \code{rlang::is_missing(x)} is \code{TRUE}, i.e., if an argument with no default is omitted from a function call.
}
\examples{
f <- function(z) {
  arg_supplied(z)
}

try(f(1)) ## No error: argument supplied
try(f())  ## Error!

# Will not throw for NULL or default arguments
try(f(NULL)) ## No error: argument supplied

f2 <- function(z = NULL) {
  arg_supplied(z)
}

try(f2()) ## No error; default provided
}
\seealso{
\code{\link[=arg_non_null]{arg_non_null()}}, \code{\link[=arg_no_NA]{arg_no_NA()}}, \code{\link[rlang:missing_arg]{rlang::is_missing()}}, \code{\link[rlang:check_required]{rlang::check_required()}}
}
