% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_unique.R
\name{arg_unique}
\alias{arg_unique}
\title{Check Unique Argument}
\usage{
arg_unique(x, ..., .arg = rlang::caller_arg(x), .msg = NULL)
}
\arguments{
\item{x}{the argument to be checked}

\item{\dots}{further argument passed to \code{\link[=anyDuplicated]{anyDuplicated()}}, which performs the checking.}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument contains only unique values.
}
\examples{
f <- function(z) {
  arg_unique(z)
}

try(f(1:3))     # No error
try(f(NULL))    # No error for NULL
try(f(c(1, 1))) # Error: repeated values
}
\seealso{
\code{\link[=anyDuplicated]{anyDuplicated()}}
}
