% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{c}
\alias{c}
\title{Argument count}
\usage{
c()
}
\value{
The number of raw arguments provided in the command line call.
}
\description{
Access the number of arguments in the argument vector.
}
\examples{
arguments <- base::c("one", "two","three") # for interactive testing
arg::v()
}
