% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{v}
\alias{v}
\title{Argument vector}
\usage{
v(..., title = "Documentation", header = "", footer = "", expose = FALSE)
}
\arguments{
\item{...}{A \code{tibble::tribble()} argument(s) and/or option(s) specification,
with \code{name}, \code{short}, \code{long}, \code{default}, \code{type} and \code{help} columns.}

\item{title}{The title for the documentation.}

\item{header}{An introductory header for the documentation.}

\item{footer}{A closing footer for the documentation.}

\item{expose}{Should the arguments and options be attached to the calling
environment instead of being returned as a named list?}
}
\value{
A named list of processed options and arguments, except if ... is
not defined, in case which the raw argument vector is returned, or if
expose is TRUE, in case which the options and arguments are attached to the
calling environment and nothing is returned.
}
\description{
Access the argument vector. If an argument(s) and/or option(s) specification
is provided, the argument vector is preprocessed first, and a named list with
arguments and options is returned instead.
}
\examples{
assignInNamespace("arguments", base::c("one", "two", "-o", "-h"), "arg") # for interactive testing
arg::v(
  ~name, ~short, ~long, ~default, ~type, ~help,
  "arg1", NA, NA, "default_arg1_value", "character", "First argument (default = \"\%default\")",
  "arg2", NA, NA, "default_arg2_value", "integer", "Second argument (default = \"\%default\")",
  "opt1", "o", "option1", NA, "logical", "Logical option",
  "opt2", "p", "option2", "Default", "character", "Character option (default = \"\%default\")",
  header = "This command line tool provides functionality to ..."
)
}
