% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_function.R
\name{arg_function}
\alias{arg_function}
\title{Check Function Argument}
\usage{
arg_function(x, .arg = rlang::caller_arg(x), .msg = NULL)
}
\arguments{
\item{x}{the argument to be checked}

\item{.arg}{the name of the argument supplied to \code{x} to appear in error messages. The default is to extract the argument's name using \code{\link[rlang:caller_arg]{rlang::caller_arg()}}. Ignored if \code{.msg} is supplied.}

\item{.msg}{an optional alternative message to display if an error is thrown instead of the default message.}
}
\value{
Returns \code{NULL} invisibly if an error is not thrown.
}
\description{
Checks whether an argument is a function.
}
\examples{
f <- function(z) {
  arg_function(z)
}

try(f(print))   # No error
try(f("print")) # Error: must be a function
}
\seealso{
\code{\link[rlang:is_function]{rlang::is_function()}}, \code{\link[=arg_is]{arg_is()}}
}
